feat: add support for filament 4.x
