--- src/Forms/Components/Actions/MediaAction.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Hugomyb\FilamentMediaAction\Forms\Components\Actions;
-
-use Filament\Forms\Components\Actions\Action;
-use Hugomyb\FilamentMediaAction\Concerns\HasMedia;
-
-class MediaAction extends Action
-{
-    use HasMedia;
-}
