--- src/Infolists/Components/Actions/MediaAction.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Hugomyb\FilamentMediaAction\Infolists\Components\Actions;
-
-use Filament\Infolists\Components\Actions\Action;
-use Hugomyb\FilamentMediaAction\Concerns\HasMedia;
-
-class MediaAction extends Action
-{
-    use HasMedia;
-}
