--- src/Tables/Actions/MediaAction.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Hugomyb\FilamentMediaAction\Tables\Actions;
-
-use Filament\Tables\Actions\Action;
-use Hugomyb\FilamentMediaAction\Concerns\HasMedia;
-
-class MediaAction extends Action
-{
-    use HasMedia;
-}
